Reviewer1: 1. Refactored TrainServer under ServerModule and lazyll-api; 2.Init TrainServer in LightEngine; 3. CLI also supported TrainServer; 4.Local and online API on LightEgine
