+++ lazyllm/module/onlineChatModule/glmModule.py
@@ -181,8 +181,10 @@ class GLMModule(OnlineChatModuleBase, FileHandlerBase):
             return 'Failed'
         elif status == 'cancelled':
             return 'Cancelled'
-        else:  # create, validating_files, queued, running
+        elif status == 'running':
             return 'Running'
+        else:  # create, validating_files, queued
+            return 'Pending'
 
     def _get_log(self, fine_tuning_job_id=None):
         if not fine_tuning_job_id and not self.fine_tuning_job_id:
